test: tests for totalSupply/borrow

+++ src/interfaces/IAllowances.sol
@@ -5,35 +5,61 @@ import "./IAsset.sol";
 
 // For full documentation refer to src/Allowances.sol";
 interface IAllowances {
-  ///////////
-  // Views //
-  ///////////
+    ///////////
+    // Views //
+    ///////////
 
-  function positiveSubIdAllowance(uint accountId, address owner, IAsset asset, uint subId, address spender)
-    external
-    view
-    returns (uint);
+    function positiveSubIdAllowance(
+        uint256 accountId,
+        address owner,
+        IAsset asset,
+        uint256 subId,
+        address spender
+    )
+        external
+        view
+        returns (uint256);
 
-  function negativeSubIdAllowance(uint accountId, address owner, IAsset asset, uint subId, address spender)
-    external
-    view
-    returns (uint);
+    function negativeSubIdAllowance(
+        uint256 accountId,
+        address owner,
+        IAsset asset,
+        uint256 subId,
+        address spender
+    )
+        external
+        view
+        returns (uint256);
 
-  function positiveAssetAllowance(uint accountId, address owner, IAsset asset, address spender)
-    external
-    view
-    returns (uint);
+    function positiveAssetAllowance(
+        uint256 accountId,
+        address owner,
+        IAsset asset,
+        address spender
+    )
+        external
+        view
+        returns (uint256);
 
-  function negativeAssetAllowance(uint accountId, address owner, IAsset asset, address spender)
-    external
-    view
-    returns (uint);
+    function negativeAssetAllowance(
+        uint256 accountId,
+        address owner,
+        IAsset asset,
+        address spender
+    )
+        external
+        view
+        returns (uint256);
 
-  ////////////
-  // Errors //
-  ////////////
+    ////////////
+    // Errors //
+    ////////////
 
-  error NotEnoughSubIdOrAssetAllowances(
-    address caller, uint accountId, int amount, uint subIdAllowance, uint assetAllowance
-  );
+    error NotEnoughSubIdOrAssetAllowances(
+        address caller,
+        uint256 accountId,
+        int256 amount,
+        uint256 subIdAllowance,
+        uint256 assetAllowance
+    );
 }
